feat: 주문 확정 기능 개발 (#13)
* test: 통합테스트를 작성한다

* feat: OrderService의 confirmOrder 메소드를 구현하고 테스트한다

* refactor: TransactionPublisher의 이름을 TransactionManager 로 변경하고 exists 메소드를 구현한다

* refactor: OrderService 가 Unit이 아닌 변경된 Order를 반환하도록 수정한다

* refactor: Order 도메인의 productquantity를 Int에서 Long으로 변경한다

* feat: OrderConfirmFacade 를 구현한다

* feat: OrderController의 confirmOrder를 구현한다

* test: 인수테스트를 수정한다

* build: 서브모듈 커밋해쉬를 업데이트한다

+++ src/main/kotlin/org/rooftop/order/domain/OrderProduct.kt
@@ -2,6 +2,6 @@ package org.rooftop.order.domain
 
 class OrderProduct(
     val productId: Long,
-    val productQuantity: Int,
+    val productQuantity: Long,
     val totalPrice: Long,
 )
